flask/requirements.txt
Flask==2.3.3
Flask-OIDC==2.4.0
requests==2.31.0
pytest==7.4.2
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7 
flask-cors==4.0.0